Some transliteration optimizations and add credo.
* Moved the majority of code under lib/cldr to better match module nesting

* Added more transliteration optimization - all {locale, number_system}
that have the same digits and symbols as {"en", "latn"} now short-circuit
transliteration since we know their digits and symbols are the same
as the ones we use for template formatting.  That includes all the
{"en-*", "latn"} combinations (there are 211 of them) and a bunch of other.

Still not happy with performance, would like to shave 100 μsec off. The
biggest areas of time use are still transliteration (80μsec if actually
transliterating) and grouping (20μsec if one group definition).