Short circuit format assembly when there is only one format part.
A compiled format is a list of tuples representing the parts of the
format.  However if there is only the format part then we don't
need to do any assembly, we just return the formatted number.

Note that negative numbers always have at least two parts
since the minus sign (or other prefix/suffix like '(' and ')'
are not part of the number format itself.

This change improves performance by up to 10 μsec.