Major refactoring of source data To package the locale information into a manner that is easier and better for distribution the locale content is now consolidated into a single file per locale.
There are additional core files to support cldr:
* available_locales.json
* plural_rules.json
* number_systems.json

Locales are no housed in the Cldr.Locale module and the
content is loaded into this module at compile time by
creating a series of `get_locale/1` functions.  All
other access to locale data is through these functions.

When requesting a locale which is configured but not
available, the Cldr.Config.get_locale/1 function will
attempt to install the locale.

For now the installation is from the `data/consolidated`
directory in the repo.  Before the next milestone
is will download from the repo the locale file on
demand.