Default json library autodetects Jason or Poison A json library needs to be configured for Cldr but since its a configuration issue we can't explicitly configure a library leading to the possibility of a library not being configured or one being configured that is not a dependency.
This commit auto-detects the presence of Jason
(preferred) or Poison as a dependency and sets
the default json_library accordingly.

The error message for when no json library is
found is also enhanced to be clear that the
error is either that there is no library or
that the configured library does not implement
the decode!/1 function.