Add grammar mapping to RBNF rule RBNF rules, especially for ordinal and spellout are different depending on locale, grammatical gender and grammatical case.
This file maps those entities to an RBNF
rule.

The default grammatical case is "nominative". The
default gender can be determined from CLDR data
(currently hosted in ex_cldr_units and needs to move)
to be hosted in ex_cldr instead).

Plural category can be determined by the number as
an argument.

Therefore it should be reasonably easy to determine the
right rule with the right answer in the default case for
many locales, and by specifying only the grammatical gender
and grammatical case where required.

This needs data to be entered and code developed and
tested to verify.