Replace `"\u00A0"` space character for regular space character en-GB
When writing tests in our application code we noticed the spaces
generated for en-GB currency in iso format isn't a regular white space
character. It's `"\u00A0"` (unicode characer escaped in Elixir). Is it
expected? I think it might not, all other ocurrences of space character
in the same file it's a regular space character.

I'm not sure if this is an error, if it should be fixed here or somewhere
else upstream. It caused confusion among my colleagues, that's why I'm
raising it up here.

What do you think? Is it something that needs to be fixed?