Silence dialyzer warnings about do_plural_rule
Libraries that depend on Cldr might use the `:extra_returns` feature in dialyzer, which will warn if the typespec of a function specifies return types that are not part of the success typing.

My understanding is that the behavior (and therefore success typing) of `do_plural_rule depends/8` on the locale used, so we cannot accurately tell what the correct return type is. Therefore, this commit posits to silence dialyzer warnings about `do_plural_rule/8`.